requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2.1
django-crispy-forms==2.4
django-guardian==3.0.0
psycopg[binary]
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
gunicorn==23.0.0
whitenoise==6.4.0
django-storages==1.14.0
python-decouple==3.8.0
